+++ clang-tools-extra/clang-tidy/bugprone/DanglingHandleCheck.cpp
@@ -18,7 +18,7 @@ using namespace clang::tidy::matchers;
 
 namespace clang {
 namespace tidy {
-namespace misc {
+namespace bugprone {
 
 namespace {
 
@@ -179,6 +179,6 @@ void DanglingHandleCheck::check(const MatchFinder::MatchResult &Result) {
       << Handle->getQualifiedNameAsString();
 }
 
-} // namespace misc
+} // namespace bugprone
 } // namespace tidy
 } // namespace clang